Adding changes from review comments.
